% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{chooseColors}
\alias{chooseColors}
\title{Choose color schema from a palette}
\usage{
chooseColors(
  palettes_name = c("Nature 10", "Light 13", "Classic 20", "Blink 23", "Hue n"),
  n_colors = 7,
  alpha = 1,
  plot_colors = FALSE
)
}
\arguments{
\item{palettes_name}{a string, the palette name, one of "Nature 10", "Light 13", "Classic 20", "Blink 23" and "Hue n", default as 'Nature 10'.}

\item{n_colors}{a positive integer, the number of colors.}

\item{alpha}{a positive real, the transparency of the color.}

\item{plot_colors}{a logical value, whether plot the selected colors.}
}
\description{
Choose color schema from a palette
}
\examples{
chooseColors()

}
