% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{drawFigs}
\alias{drawFigs}
\title{Draw a figure using a group of ggplot objects}
\usage{
drawFigs(
  pList,
  layout.dim = NULL,
  common.legend = FALSE,
  legend.position = "right",
  ...
)
}
\arguments{
\item{pList}{a list with component ggplot objects.}

\item{layout.dim}{a integer vector with length 2, the layout of subplots in rows and columns.}

\item{common.legend}{a logical value, whether use common legend for all subplots.}

\item{legend.position}{a string, the position of legend.}

\item{...}{other arguments that pass to \code{\link{ggarrange}}.}
}
\value{
return a new ggplot object.
}
\description{
Draw a figure using a group of ggplot objects
}
