\name{firstup}
\alias{firstup}
\title{Set the first letter of a string vector to captial}
\description{
  Set the first letter of a string vector to captial.
}
\usage{
  firstup(x)
}
\arguments{
  \item{x}{a string vector.}
}
\details{
  Nothing
}
\value{
  Return a string vector with first letter capital.
}

\author{
  Wei Liu
}
\note{
  nothing
}
\seealso{
  None
}
\examples{
  x <- c("good", "Morning")
  firstup(x)
}