\name{plot_RGB}
\alias{plot_RGB}
\title{Spatial RGB heatmap}
\description{
  Plot spatial RGB heatmap.
}
\usage{
  plot_RGB(position, embed_3d, pointsize=2,textsize=15)
}
\arguments{
  \item{position}{a coordinates matrix with two columns: x-coordinate and y-coordinate.}
  \item{embed_3d}{a embedding matrix with three columns: x, y and z embeddings.}
  \item{pointsize}{the size of point in the scatter plot.}
  \item{textsize}{the text size in the plot.}
}
\details{
  Nothing
}
\value{
  Return a ggplot2 object.
}

\author{
  Wei Liu
}
\note{
  nothing
}
\seealso{
  None
}
\examples{

}